+++ src/lambdas/ingestProject/lambdaFetchRootSpans/extractPageInfo.ts
