+++ src/lambdas/ingestProject/lambdaFetchRootSpans/fetchRootSpans.ts
