+++ src/lambdas/ingestProject/lambdaFetchRootSpans/formatRootSpans.ts
