+++ src/lambdas/ingestProject/lambdaRDS/getLastRootSpan.ts
