+++ src/lambdas/ingestProject/lambdaRDS/insertRootSpans.ts
